__legacy/app/requirements.txt
amqp==5.2.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confusable-homoglyphs==3.3.1
crispy-bootstrap5==2024.2
Django==5.0.3
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-redis==5.4.0
django-registration==3.4
djangorestframework==3.15.1
dnspython==2.6.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
kombu==5.3.6
moexalgo==1.0.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
prompt-toolkit==3.0.43
psycopg==3.1.18
psycopg-binary==3.1.18
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
